requirements.txt
chromadb==0.4.18
llama-index-llms-openai==0.2.1
llama-index-embeddings-openai==0.1.5
llama-index-vector-stores-chroma==0.1.3
llama-index-core==0.10.1
